Move the implementation of a `DragAndDropManager` delegating to `PlatformDragAndDropManager` to PlatformDragAndDropManager.kt
